pyproject.toml
#####################################################################################
# A python System Verilog Parser and AST
# Copyright (C) 2024  RISCY-Lib Contributors
#
# This library is free software; you can redistribute it and/or
# modify it under the terms of the GNU Lesser General Public
# License as published by the Free Software Foundation; either
# version 2.1 of the License, or (at your option) any later version.
#
# This library is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
# Lesser General Public License for more details.
#
# You should have received a copy of the GNU Lesser General Public
# License along with this library; If not, see <https://www.gnu.org/licenses/>.
#####################################################################################

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sverilogpy"
dynamic = ["version"]
description = "A python System Verilog Parser and AST"
readme = "README.md"
authors = [{name = "Benjamin Davis", email = "allrisc.dev@gmail.com"}]
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "License :: OSI Approved :: GNU Lesser General Public License v2 or later (LGPLv2+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]
keywords = ["Verilog", "tool", "SystemVerilog"]

dependencies = ["antlr4-python3-runtime", "attrs"]
requires-python = ">=3.8"

[project.urls]
Homepage = "https://github.com/RISCY-Lib/sverilogpy"
Documentation = "https://github.com/RISCY-Lib/sverilogpy/blob/main/README.md"

[project.optional-dependencies]
dev = ["flake8", "mypy", "tox", "pytest", "antlr4-tools", "setuptools"]